done db creation

+++ src/main/java/ecommerce/api/entity/base/EntityBase.java
@@ -28,7 +28,7 @@ public abstract class EntityBase {
 
     //in case we exceed the limit of 6 digits, we will use the base32 encoding
     @Column(name = "code",insertable = false, updatable = false)
-    protected Integer code;
+    protected Long code;
 
     public EntityBase() {
         this.id = UUID.randomUUID();
